remove trigger digest email endpoint instead allow it to be triggered via console
